add database

+++ WeatherApp/Models/GeoInformationModel.cs
@@ -5,9 +5,9 @@ namespace WeatherApp.Models
     public class GeoInformationModel
     {
         [JsonPropertyName("lat")]
-        public decimal Lat { get; set; }
+        public double Lat { get; set; }
 
         [JsonPropertyName("lon")]
-        public decimal Lon { get; set; }
+        public double Lon { get; set; }
     }
 }
